+++ stroom-pipeline/src/main/java/stroom/pipeline/server/writer/RollingStreamAppender.java
@@ -54,7 +54,7 @@ import java.io.IOException;
         PipelineElementType.VISABILITY_STEPPING}, icon = ElementIcons.STREAM)
 public class RollingStreamAppender extends AbstractRollingAppender implements RollingDestinationFactory {
     private static final int MB = 1024 * 1024;
-    private static final int DEFAULT_MAX_SIZE = 100 * MB;
+    private static final int DEFAULT_ROLL_SIZE = 100 * MB;
 
     private static final int SECOND = 1000;
     private static final int MINUTE = 60 * SECOND;
@@ -70,7 +70,7 @@ public class RollingStreamAppender extends AbstractRollingAppender implements Ro
     private String streamType;
     private boolean segmentOutput = true;
     private long frequency = HOUR;
-    private long maxSize = DEFAULT_MAX_SIZE;
+    private long rollSize = DEFAULT_ROLL_SIZE;
 
     private boolean validatedSettings;
 
@@ -102,7 +102,7 @@ public class RollingStreamAppender extends AbstractRollingAppender implements Ro
 
         final String nodeName = nodeCache.getDefaultNode().getName();
         final StreamTarget streamTarget = streamStore.openStreamTarget(stream);
-        return new RollingStreamDestination(key, frequency, maxSize, streamStore,
+        return new RollingStreamDestination(key, frequency, rollSize, streamStore,
                 streamTarget, nodeName, System.currentTimeMillis());
     }
 
@@ -133,8 +133,8 @@ public class RollingStreamAppender extends AbstractRollingAppender implements Ro
                 throw new ProcessException("Rolling frequency must be greater than 0");
             }
 
-            if (maxSize <= 0) {
-                throw new ProcessException("Max size must be greater than 0");
+            if (rollSize <= 0) {
+                throw new ProcessException("Roll size must be greater than 0");
             }
         }
     }
@@ -171,17 +171,17 @@ public class RollingStreamAppender extends AbstractRollingAppender implements Ro
     }
 
     @PipelineProperty(description = "Choose the maximum size that a stream can be before it is rolled.", defaultValue = "100M")
-    public void setMaxSize(final String maxSize) {
-        if (maxSize != null && maxSize.trim().length() > 0) {
+    public void setRollSize(final String rollSize) {
+        if (rollSize != null && rollSize.trim().length() > 0) {
             try {
-                final Long value = ModelStringUtil.parseIECByteSizeString(maxSize);
+                final Long value = ModelStringUtil.parseIECByteSizeString(rollSize);
                 if (value == null) {
-                    throw new PipelineFactoryException("Incorrect value for max size: " + maxSize);
+                    throw new PipelineFactoryException("Incorrect value for max size: " + rollSize);
                 }
 
-                this.maxSize = value;
+                this.rollSize = value;
             } catch (final NumberFormatException e) {
-                throw new PipelineFactoryException("Incorrect value for max size: " + maxSize);
+                throw new PipelineFactoryException("Incorrect value for max size: " + rollSize);
             }
         }
     }